Rename image.rs to image_gc_settings.rs
