+++ edgelet/edgelet-settings/src/base/image_gc_settings.rs
