+++ edgelet/edgelet-settings/src/base/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 
 pub mod aziot;
-pub mod image;
+pub mod image_gc_settings;
 pub mod module;
 pub mod uri;
 pub mod watchdog;
@@ -39,7 +39,7 @@ pub trait RuntimeSettings {
 
     fn additional_info(&self) -> &std::collections::BTreeMap<String, String>;
 
-    fn image_garbage_collection(&self) -> &image::Settings;
+    fn image_garbage_collection(&self) -> &image_gc_settings::Settings;
 }
 
 #[derive(Clone, Debug, Default, Eq, PartialEq, serde::Deserialize, serde::Serialize)]
@@ -128,8 +128,11 @@ pub struct Settings<ModuleConfig> {
     #[serde(default, skip_serializing_if = "std::collections::BTreeMap::is_empty")]
     pub additional_info: std::collections::BTreeMap<String, String>,
 
-    #[serde(default, skip_serializing_if = "image::Settings::is_default")]
-    pub image_garbage_collection: image::Settings,
+    #[serde(
+        default,
+        skip_serializing_if = "image_gc_settings::Settings::is_default"
+    )]
+    pub image_garbage_collection: image_gc_settings::Settings,
 }
 
 pub(crate) fn default_allow_elevated_docker_permissions() -> bool {
@@ -211,7 +214,7 @@ impl<T: Clone> RuntimeSettings for Settings<T> {
         &self.additional_info
     }
 
-    fn image_garbage_collection(&self) -> &image::Settings {
+    fn image_garbage_collection(&self) -> &image_gc_settings::Settings {
         &self.image_garbage_collection
     }
 }
