+++ edgelet/iotedge/src/config/import/mod.rs
@@ -14,7 +14,7 @@ mod old_config;
 
 use std::path::{Path, PathBuf};
 
-use edgelet_settings::base::image::Settings;
+use edgelet_settings::base::image_gc_settings::Settings;
 use edgelet_utils::YamlFileSource;
 
 use aziotctl_common::config as common_config;
